+++ crates/hlight/src/syntax.rs
@@ -138,6 +138,8 @@ pub fn find_syntax_name<'a>(set: &'a SyntaxSet, name: &str) -> &'a SyntaxReferen
 
 #[cfg(test)]
 mod tests {
+  use std::borrow::Cow;
+
   use crate::HighlightResource;
 
   #[test]
@@ -155,5 +157,46 @@ mod tests {
                     ext:{ext:?}\n---"
         )
       });
+  }
+
+  #[ignore]
+  #[test]
+  fn test_load_custom_syntax_set() {
+    use std::sync::OnceLock;
+
+    // use hlight::
+    use crate::{
+      HighlightResource,
+      syntax::{SyntaxSet, load_syntax_set},
+    };
+
+    const SYNTAXES: &[u8] = include_bytes!(concat!(
+      env!("CARGO_MANIFEST_DIR"),
+      "/../hlight-assets",
+      "/assets/set/syntax.packdump"
+    ));
+
+    fn static_syntax_set() -> &'static SyntaxSet {
+      static S: OnceLock<SyntaxSet> = OnceLock::new();
+      S.get_or_init(|| load_syntax_set(Some(SYNTAXES)))
+    }
+
+    fn show_syntax_set(set: &SyntaxSet) {
+      for (name, ext) in set
+        .syntaxes()
+        .iter()
+        .map(|x| (&x.name, &x.file_extensions))
+      {
+        println!(
+          "name: {name}\n\
+          ext: {ext:?}\n---"
+        )
+      }
+    }
+
+    let res = HighlightResource::default()
+      .with_syntax_set(Cow::Borrowed(static_syntax_set()));
+
+    show_syntax_set(res.get_syntax_set())
   }
 }